@@ -9,19 +9,62 @@
         }
     ],
     "require": {
-        "silex/silex": "~2.0",
-        "doctrine/dbal": "~2.5",
+        "php" : "^7.4",
+        "silex/silex": "2.0.*",
+        "monolog/monolog": "^1.17",
+        "doctrine/dbal": "2.5.*",
         "twig/twig": "~1.28",
         "symfony/twig-bridge": "~2.8|3.0.*",
         "symfony/asset": "~2.8|3.0.*",
-        "symfony/monolog-bridge": "~2.8|3.0.*"        
+        "symfony/monolog-bridge": "~2.8|3.0.*"   ,
+        "symfony/http-foundation": "^3.4"     
+    },
+    "require-dev": {
+        "friendsofphp/php-cs-fixer": "*",
+        "phpmd/phpmd": "^2.15",
+        "phpstan/phpstan": "^1.10",
+        "squizlabs/php_codesniffer": "^3.9",
+        "vimeo/psalm": "*"
     },
     "autoload": {
         "psr-4": {
             "Semeformation\\Mvc\\Cinema_crud\\": "src/"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "Semeformation\\Mvc\\Cinema_crud\\Test\\": "tests/"
+        }
+    },
     "config": {
-        "vendor-dir": "src/vendor"
+        "vendor-dir": "src/vendor",
+        "process-timeout": 0,
+        "sort-packages": true,
+        "allow-plugins": {
+            "composer/package-versions-deprecated": true
+        }
+    },
+    "scripts": {
+        "phpmd" : "phpmd --verbose --exclude *\\vendor\\* --color src\\ text phpmd_config.xml",
+        "cs:check": [
+            "@putenv PHP_CS_FIXER_IGNORE_ENV=1",
+            "php-cs-fixer fix --dry-run --format=txt --verbose --diff --config=.php-cs-fixer.php --ansi"
+        ],
+        "cs:fix": [
+            "@putenv PHP_CS_FIXER_IGNORE_ENV=1",
+            "php-cs-fixer fix --config=.php-cs-fixer.php --ansi --verbose"
+        ],
+        "sniffer:check": "phpcs --standard=.phpcs.xml --report=full --report-width=120 --colors -p -s",
+        "sniffer:fix": "phpcbf --standard=.phpcs.xml -p",
+        "stan": "phpstan analyse -c phpstan.neon --ansi",
+        "stan:bl": "phpstan analyse -c phpstan.neon --ansi --generate-baseline",
+        "psalm:init" : "./src/vendor/bin/psalm --init",
+        "psalm" : "./src/vendor/bin/psalm",
+        "start": "php -S localhost:8080 -t src/",
+        "test:all": [
+            "@cs:check",
+            "@sniffer:check",
+            "@stan"
+        ]
     }
 }
