@@ -9,17 +9,19 @@ use Symfony\Component\HttpFoundation\Request;
  *
  * @author Seme
  */
-class Controller {
-
+class Controller
+{
     /**
      * Se charge d'extraire les données POST de la requête sous forme de tableau
      * @param Request $request Requête HTTP POST
      * @param array $variables Tableau de noms de variables à récupérer
      * @return array Tableau de variables extraites de la requête POST
      */
-    protected function extractArrayFromPostRequest(Request $request,
-            array $variables): array {
-        $entries = array();
+    protected function extractArrayFromPostRequest(
+        Request $request,
+        array $variables
+    ): array {
+        $entries = [];
         // boucle de parcours des variables à extraire
         foreach ($variables as $variable) {
             $entries[$variable] = $request->request->get($variable);
@@ -33,9 +35,11 @@ class Controller {
      * @param array $variables Tableau de noms de variables à récupérer
      * @return array Tableau de variables extraites de la requête POST
      */
-    protected function extractArrayFromGetRequest(Request $request,
-            array $variables): array {
-        $entries = array();
+    protected function extractArrayFromGetRequest(
+        Request $request,
+        array $variables
+    ): array {
+        $entries = [];
         // boucle de parcours des variables à extraire
         foreach ($variables as $variable) {
             $entries[$variable] = $request->query->get($variable);
@@ -43,4 +47,35 @@ class Controller {
         return $entries;
     }
 
+    protected function redirectIfUserNotConnected($request, $app)
+    {
+        if ($app['session']->get('user') === null) {
+            // renvoi à la page d'accueil
+            return $app->redirect($request->getBasePath() . '/home');
+        }
+    }
+
+    protected function redirectIfUserNotConnectedOrNotAdmin($request, $app)
+    {
+        if ($app['session']->get('user') === null || $app['session']->get('user')['username'] !== 'admin@adm.adm') {
+            // renvoi à la page d'accueil
+            return $app->redirect($request->getBasePath() . '/home');
+        }
+    }
+
+    protected function checkIfUserIsConnected($app)
+    {
+        if ($app['session']->get('user') !== null) {
+            return true;
+        }
+        return false;
+    }
+
+    protected function checkIfUserIsConnectedAndAdmin($app)
+    {
+        if ($app['session']->get('user') !== null && $app['session']->get('user')['username'] === 'admin@adm.adm') {
+            return true;
+        }
+        return false;
+    }
 }
