@@ -11,15 +11,16 @@ use Semeformation\Mvc\Cinema_crud\exceptions\BusinessObjectDoNotExist;
  *
  * @author User
  */
-class CinemaDAO extends DAO {
-
+class CinemaDAO extends DAO
+{
     /**
      * Crée un cinéma à partir d'une ligne de la BDD.
      *
      * @param array $row La ligne de résultat de la BDD.
      * @return Cinema
      */
-    protected function buildBusinessObject($row) {
+    protected function buildBusinessObject($row)
+    {
         $cinema = new Cinema();
         $cinema->setCinemaId($row['CINEMAID']);
         $cinema->setDenomination($row['DENOMINATION']);
@@ -29,19 +30,20 @@ class CinemaDAO extends DAO {
 
     /**
      * Recherche un cinéma à partir de son identifiant
-     * @param string $cinemaID
+     * @param string $id
      * @return Cinema
-     * @throws Exception
+     * @throws \Exception
      */
-    public function find(...$cinemaID) {
-        $requete  = "SELECT * FROM cinema WHERE cinemaID = ?";
-        $resultat = $this->getDb()->fetchAssoc($requete, [$cinemaID[0]]);
+    public function find(...$id)
+    {
+        $requete = "SELECT * FROM cinema WHERE cinemaID = ?";
+        $resultat = $this->getDb()->fetchAssoc($requete, [$id[0]]);
         // si trouvé
-        if ($resultat) {
+        if ($resultat !== false) {
             // on crée et on retourne l'objet métier Cinema
             return $this->buildBusinessObject($resultat);
         } else {
-            throw new BusinessObjectDoNotExist('Aucun cinéma trouvé pour l\'id=' . $cinemaID[0]);
+            throw new BusinessObjectDoNotExist('Aucun cinéma trouvé pour l\'id=' . $id[0]);
         }
     }
 
@@ -49,9 +51,10 @@ class CinemaDAO extends DAO {
      * Recherche tous les cinémas en BDD et retourne le résultat sous forme de tableau
      * @return array Le tableau de cinémas
      */
-    public function findAll() {
+    public function findAll()
+    {
         // requête d'extraction de tous les cinémas
-        $sql       = "SELECT * FROM cinema ORDER BY denomination ASC";
+        $sql = "SELECT * FROM cinema ORDER BY denomination ASC";
         $resultats = $this->getDb()->fetchAll($sql);
 
         // on extrait les objets métiers des résultats
@@ -61,11 +64,13 @@ class CinemaDAO extends DAO {
     /**
      * Renvoie la liste des cinémas d'un film
      * @param integer $filmID
-     * @return array Le tableau d'objets Cinema
+     *
+     * @return array<object>|null Le tableau d'objets Cinema
      */
-    public function findAllByFilmId($filmID) {
+    public function findAllByFilmId($filmID)
+    {
         // requête qui nous permet de récupérer la liste des cinémas pour un film donné
-        $requete   = "SELECT DISTINCT c.* FROM cinema c"
+        $requete = "SELECT DISTINCT c.* FROM cinema c"
                 . " INNER JOIN seance s ON c.cinemaID = s.cinemaID"
                 . " AND s.filmID = " . $filmID;
         // on extrait les résultats
@@ -76,13 +81,16 @@ class CinemaDAO extends DAO {
 
     /**
      * Renvoie une liste de cinémas qui ne projettent pas le film donné
+     *
      * @param integer $filmID
-     * @return array Le tableau d'objets Cinema
+     *
+     * @return array<object>|null Le tableau d'objets Cinema
      */
-    public function findAllByFilmIdNotIn($filmID) {
+    public function findAllByFilmIdNotIn($filmID)
+    {
         // requête de récupération des titres et des identifiants des films
         // qui n'ont pas encore été programmés dans ce cinéma
-        $requete   = "SELECT c.CINEMAID, c.DENOMINATION, c.ADRESSE "
+        $requete = "SELECT c.CINEMAID, c.DENOMINATION, c.ADRESSE "
                 . "FROM cinema c"
                 . " WHERE c.cinemaID NOT IN ("
                 . "SELECT cinemaID"
@@ -96,21 +104,43 @@ class CinemaDAO extends DAO {
     }
 
     /**
+     * Renvoie une liste de cinémas qui projettent au moins un film
+     *
+     * @return array<object>|null Le tableau d'objets Cinema
+     */
+    public function findAllOnAir()
+    {
+        $resultatsFormatted = [];
+        $requete = "SELECT DISTINCT CINEMAID FROM seance";
+        // extraction des résultats
+        $resultats = $this->getDb()->fetchAll($requete);
+        foreach ($resultats as $row) {
+            $resultatsFormatted[] = $row['CINEMAID'];
+        }
+        // retour du résultat
+        return $resultatsFormatted;
+    }
+
+    /**
      * Sauvegarde un objet Cinema en BDD
      * @param Cinema $cinema
      */
-    public function save(Cinema $cinema) {
+    public function save(Cinema $cinema)
+    {
         // je récupère les données du cinéma sous forme de tableau
-        $donneesCinema = array(
+        $donneesCinema = [
             'denomination' => $cinema->getDenomination(),
             'adresse'      => $cinema->getAdresse(),
-        );
+        ];
 
         // Si le cinéma existe déja
-        if ($cinema->getCinemaId()) {
+        if ($cinema->getCinemaId() !== null) {
             // il faut faire une mise à jour
-            $this->getDb()->update('cinema', $donneesCinema,
-                    array('cinemaId' => $cinema->getCinemaId()));
+            $this->getDb()->update(
+                'cinema',
+                $donneesCinema,
+                ['cinemaId' => $cinema->getCinemaId()]
+            );
         } else {
             // Sinon, nous faisons une insertion
             $this->getDb()->insert('cinema', $donneesCinema);
@@ -123,11 +153,12 @@ class CinemaDAO extends DAO {
 
     /**
      * Supprime un cinéma
+     *
      * @param integer $cinemaID
      */
-    public function delete($cinemaID) {
+    public function delete($cinemaID)
+    {
         // Supprime le cinéma
-        $this->getDb()->delete('cinema', array('cinemaId' => $cinemaID));
+        $this->getDb()->delete('cinema', ['cinemaId' => $cinemaID]);
     }
-
 }
