@@ -11,19 +11,20 @@ use Semeformation\Mvc\Cinema_crud\exceptions\BusinessObjectDoNotExist;
  *
  * @author User
  */
-class FilmDAO extends DAO {
-
+class FilmDAO extends DAO
+{
     /**
      * Crée un film à partir d'une ligne de la BDD.
      *
      * @param array $row La ligne de résultat de la BDD.
      * @return Film
      */
-    protected function buildBusinessObject($row) {
+    protected function buildBusinessObject($row)
+    {
         $film = new Film();
         $film->setFilmId($row['FILMID']);
         $film->setTitre($row['TITRE']);
-        if (array_key_exists('TITREORIGINAL', $row)) {
+        if (array_key_exists('TITREORIGINAL', $row) === true) {
             $film->setTitreOriginal($row['TITREORIGINAL']);
         }
         return $film;
@@ -31,20 +32,23 @@ class FilmDAO extends DAO {
 
     /**
      * Retourne le BO Film en fonction de son identifiant
-     * @param type $filmId
-     * @return type
-     * @throws Exception
+     * @param array<string> $id
+     * @return Film
+     * @throws \Exception
      */
-    public function find(...$filmId) {
-        $requete  = "SELECT * FROM film WHERE filmID = ?";
-        $resultat = $this->getDb()->fetchAssoc($requete, [
-            $filmId[0]]);
+    public function find(...$id)
+    {
+        $requete = "SELECT * FROM film WHERE filmID = ?";
+        $resultat = $this->getDb()->fetchAssoc(
+            $requete,
+            [$id[0]]
+        );
         // si trouvé
-        if ($resultat) {
+        if ($resultat !== false) {
             // on récupère et on retourne l'objet Film
             return $this->buildBusinessObject($resultat);
         } else {
-            throw new BusinessObjectDoNotExist('Aucun film trouvé pour l\'id=' . $filmId[0]);
+            throw new BusinessObjectDoNotExist("Aucun film trouvé pour l'id=" . $id[0]);
         }
     }
 
@@ -52,9 +56,10 @@ class FilmDAO extends DAO {
      * Retourne tous les films de la base de données
      * @return array
      */
-    public function findAll() {
+    public function findAll()
+    {
         // requête d'extraction de tous les films
-        $sql       = "SELECT * FROM film ORDER BY titre ASC";
+        $sql = "SELECT * FROM film ORDER BY titre ASC";
         $resultats = $this->getDb()->fetchAll($sql);
 
         // on extrait les objets métiers des résultats
@@ -63,18 +68,20 @@ class FilmDAO extends DAO {
 
     /**
      * Retourne les films d'un cinéma
-     * @param type $cinemaID
+     * @param string $cinemaID
      * @return array Tableau d'objes Film
      */
-    public function findAllByCinemaId($cinemaID) {
+    public function findAllByCinemaId($cinemaID)
+    {
         // requête qui nous permet de récupérer la liste des films pour un cinéma donné
-        $requete   = "SELECT DISTINCT f.* FROM film f"
+        $requete = "SELECT DISTINCT f.* FROM film f"
                 . " INNER JOIN seance s ON f.filmID = s.filmID"
                 . " AND s.cinemaID = :cinemaID";
         // on extrait les résultats
-        $resultats = $this->getDb()->fetchAll($requete,
-                [
-            'cinemaID' => $cinemaID]);
+        $resultats = $this->getDb()->fetchAll(
+            $requete,
+            ['cinemaID' => $cinemaID]
+        );
         // on extrait les objets métiers des résultats
         return $this->extractObjects($resultats);
     }
@@ -83,12 +90,14 @@ class FilmDAO extends DAO {
      * Méthode qui ne renvoie que les films non encore marqués
      * comme favoris par l'utilisateur passé en paramètre
      * @param int $userID Identifiant de l'utilisateur
-     * @return Film[] Films présents dans la base respectant les critÃ¨res
+     *
+     * @return array<object> Films présents dans la base respectant les critères
      */
-    public function findAllByUserIdNotIn($userID) {
+    public function findAllByUserIdNotIn($userID): array
+    {
         // requête de récupération des titres et des identifiants des films
         // qui n'ont pas encore été marqués comme favoris par l'utilisateur
-        $requete   = "SELECT f.FILMID, f.TITRE "
+        $requete = "SELECT f.FILMID, f.TITRE "
                 . "FROM film f"
                 . " WHERE f.filmID NOT IN ("
                 . "SELECT filmID"
@@ -97,21 +106,26 @@ class FilmDAO extends DAO {
                 . ")"
                 . " ORDER BY f.TITRE ASC";
         // extraction de résultat
-        $resultats = $this->getDb()->fetchAll($requete, [
-            'id' => $userID]);
+        $resultats = $this->getDb()->fetchAll(
+            $requete,
+            ['id' => $userID]
+        );
         // on extrait les objets métiers des résultats
         return $this->extractObjects($resultats);
     }
 
     /**
      * Renvoie une liste de films pas encore programmés pour un cinema donné
+     *
      * @param integer $cinemaID
-     * @return array
+     *
+     * @return array<object>
      */
-    public function findAllByCinemaIdNotIn($cinemaID) {
+    public function findAllByCinemaIdNotIn($cinemaID): array
+    {
         // requête de récupération des titres et des identifiants des films
         // qui n'ont pas encore été programmés dans ce cinéma
-        $requete  = "SELECT f.FILMID, f.TITRE "
+        $requete = "SELECT f.FILMID, f.TITRE "
                 . "FROM film f"
                 . " WHERE f.filmID NOT IN ("
                 . "SELECT filmID"
@@ -119,29 +133,52 @@ class FilmDAO extends DAO {
                 . " WHERE cinemaID = :id"
                 . ")";
         // extraction de résultat
-        $resultat = $this->getDb()->fetchAll($requete, [
-            'id' => $cinemaID]);
+        $resultat = $this->getDb()->fetchAll(
+            $requete,
+            ['id' => $cinemaID]
+        );
         // on extrait les objets métiers des résultats
         return $this->extractObjects($resultat);
+    }
+
+    /**
+     * Renvoie une liste de films qui sont projetés dans au moins un cinéma
+     *
+     * @return array<object>|null Le tableau d'objets Film
+     */
+    public function findAllOnAir()
+    {
+        $resultatsFormatted = [];
+        $requete = "SELECT DISTINCT FILMID FROM seance";
+        $resultats = $this->getDb()->fetchAll($requete);
+        foreach ($resultats as $row) {
+            $resultatsFormatted[] = $row['FILMID'];
+        }
+        return $resultatsFormatted;
     }
 
     /**
      * Sauvegarde un objet Film en BDD
      * @param Film $film
      */
-    public function save(Film $film) {
+    public function save(Film $film)
+    {
         // je récupère les données du film sous forme de tableau
-        $donneesFilm = array(
+        $donneesFilm = [
             'titre'         => $film->getTitre(),
             'titreOriginal' => $film->getTitreOriginal(),
-        );
+        ];
 
         // Si le film existe déja
-        if ($film->getFilmId()) {
+        if ($film->getFilmId() !== null) {
             // il faut faire une mise à jour
-            $this->getDb()->update('film', $donneesFilm,
-                    array(
-                'filmId' => $film->getFilmId()));
+            $this->getDb()->update(
+                'film',
+                $donneesFilm,
+                [
+                    'filmId' => $film->getFilmId(),
+                ]
+            );
         } else {
             // Sinon, nous faisons une insertion
             $this->getDb()->insert('film', $donneesFilm);
@@ -154,11 +191,14 @@ class FilmDAO extends DAO {
 
     /**
      * Supprime un film
+     *
      * @param integer $movieID
      */
-    public function delete($movieID) {
-        $this->getDb()->delete('film', array(
-            'filmId' => $movieID));
+    public function delete($movieID)
+    {
+        $this->getDb()->delete(
+            'film',
+            ['filmId' => $movieID]
+        );
     }
-
 }
