@@ -5,17 +5,17 @@ namespace Semeformation\Mvc\Cinema_crud\includes;
 use Doctrine\DBAL\Connection;
 use Psr\Log\LoggerInterface;
 
-abstract class DAO {
-
+abstract class DAO
+{
     /**
      * Connexion à la BDD
-     * @var Doctrine\DBAL\Connection 
+     * @var Connection
      */
     private $db;
 
     /**
      * Logger du DAO
-     * @var Psr\Log\LoggerInterface 
+     * @var LoggerInterface
      */
     protected $logger;
 
@@ -24,10 +24,12 @@ abstract class DAO {
      * @param Connection $connexion
      * @param LoggerInterface $logger
      */
-    public function __construct(Connection $connexion = null,
-            LoggerInterface $logger = null) {
+    public function __construct(
+        Connection $connexion = null,
+        LoggerInterface $logger = null
+    ) {
         // init. de la connexion à la BDD
-        $this->db     = $connexion;
+        $this->db = $connexion;
         // init. du logger
         $this->logger = $logger;
     }
@@ -36,15 +38,17 @@ abstract class DAO {
      * Donne accès à la connexion à la BDD
      * @return Connection
      */
-    protected function getDb() {
+    protected function getDb()
+    {
         return $this->db;
     }
 
     /**
      * Donne accès au logger du DAO
-     * @return Psr\Log\LoggerInterface
+     * @return LoggerInterface
      */
-    public function getLogger() {
+    public function getLogger()
+    {
         return $this->logger;
     }
 
@@ -52,25 +56,26 @@ abstract class DAO {
      * Méthode abstraite de construction d'un objet métier à partir d'une ligne de la BDD
      * Cette méthode DOIT être redéfinie dans les classes filles
      */
-    protected abstract function buildBusinessObject($row);
+    abstract protected function buildBusinessObject($row);
 
     /**
      * Méthode abstraite de recherche d'un BO à partir de son id
      */
-    public abstract function find(...$id);
+    abstract public function find(...$id);
 
     /**
      * Recherche tous les BO présents dans la BDD
      */
-    public abstract function findAll();
+    abstract public function findAll();
 
     /**
      * Construit un tableau d'objets métiers à partir d'un résultat de BDD
-     * @param type $rows
-     * @return array
+     * @param array $rows
+     * @return array<Object>
      */
-    protected function buildBusinessObjects($rows) {
-        $objets = array();
+    protected function buildBusinessObjects($rows)
+    {
+        $objets = [];
         foreach ($rows as $row) {
             $objets[] = $this->buildBusinessObject($row);
         }
@@ -79,17 +84,18 @@ abstract class DAO {
 
     /**
      * Extrait des résultats, un vecteur d'objets métiers
-     * @param type $results
-     * @return Object[]
+     * @param array|null $results
+     * @return array<object>
      */
-    protected function extractObjects($results) {
-        if (!is_null($results)) {
+    protected function extractObjects($results): array
+    {
+        if ($results !== null) {
             // on crée les objets métiers
             $objects = $this->buildBusinessObjects($results);
             // on retourne le résultat
             return $objects;
         } else {
-            return array();
+            return [];
         }
     }
 
