@@ -7,14 +7,15 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\RedirectResponse;
 use Silex\Application;
 use DateTime;
+use Semeformation\Mvc\Cinema_crud\models\Seance;
 
 /**
  * Description of ShowtimesController
  *
  * @author User
  */
-class ShowtimesController extends Controller {
-
+class ShowtimesController extends Controller
+{
     /**
      * Route liste des séances d'un film
      * @param string $filmId
@@ -22,37 +23,28 @@ class ShowtimesController extends Controller {
      * @param Application $app
      * @return string
      */
-    public function movieShowtimes(string $filmId, Request $request = null,
-            Application $app = null) {
-        $adminConnected = false;
+    public function movieShowtimes(
+        string $filmId = null,
+        Request $request = null,
+        Application $app = null
+    ) {
+        $adminConnected = $this->checkIfUserIsConnectedAndAdmin($app);
 
-        // si l'utilisateur est pas connecté et qu'il est amdinistrateur
-        if ($app['session']->get('user') && $app['session']->get('user')['username'] ==
-                'admin@adm.adm') {
-            $adminConnected = true;
-        }
-
-        // on assainit les entrées
-        $entries['filmID'] = $filmId;
-        // si l'identifiant du film a bien été passé en GET'
-        if ($entries && !is_null($entries['filmID']) && $entries['filmID'] !== '') {
-            // on récupère l'identifiant du cinéma
-            $filmID = $entries['filmID'];
+        // si l'identifiant du film a bien été passé en GET
+        if ($filmId !== null && $filmId !== "") {
             // puis on récupère les informations du film en question
-            $film   = $app['dao.seance']->getFilmDAO()->find($filmID);
+            $film = $app['dao.seance']->getFilmDAO()->find($filmId);
 
             // on récupère les cinémas qui ne projettent pas encore le film
-            $cinemasUnplanned = $app['dao.seance']->getCinemaDAO()->findAllByFilmIdNotIn($filmID);
-        }
-        // sinon, on retourne à l'accueil
-        else {
+            $cinemasUnplanned = $app['dao.seance']->getCinemaDAO()->findAllByFilmIdNotIn($filmId);
+        } else {
             // renvoi à la page d'accueil
             return $app->redirect($request->getBasePath() . '/home');
         }
 
         // on récupère la liste des cinémas de ce film
-        $cinemas = $app['dao.seance']->getCinemaDAO()->findAllByFilmId($filmID);
-        $seances = $app['dao.seance']->findAllByFilmId($cinemas, $filmID);
+        $cinemas = $app['dao.seance']->getCinemaDAO()->findAllByFilmId($filmId);
+        $seances = $app['dao.seance']->findAllByFilmId($cinemas, $filmId);
 
         // données de la vue
         $donnees = [
@@ -73,39 +65,30 @@ class ShowtimesController extends Controller {
      * @param string $cinemaId
      * @return string
      */
-    public function cinemaShowtimes(Request $request = null,
-            Application $app = null, string $cinemaId = null) {
-        $adminConnected = false;
-
+    public function cinemaShowtimes(
+        Request $request = null,
+        Application $app = null,
+        string $cinemaId = null
+    ) {
         // si l'utilisateur est pas connecté et qu'il est amdinistrateur
-        if ($app['session']->get('user') && $app['session']->get('user')['username'] ==
-                'admin@adm.adm') {
-            $adminConnected = true;
-        }
-
-        // on assainit les entrées
-        $entries['cinemaID'] = $cinemaId;
+        $adminConnected = $this->checkIfUserIsConnectedAndAdmin($app);
 
         // si l'identifiant du cinéma a bien été passé en GET
-        if ($entries && !is_null($entries) && $entries['cinemaID'] != '') {
-            // on récupère l'identifiant du cinéma
-            $cinemaID = $entries['cinemaID'];
+        if ($cinemaId !== null && $cinemaId !== "") {
             // puis on récupère les informations du cinéma en question
-            $cinema   = $app['dao.seance']->getCinemaDAO()->find($cinemaID);
+            $cinema = $app['dao.seance']->getCinemaDAO()->find($cinemaId);
 
             // on récupère les films pas encore projetés
-            $filmsUnplanned = $app['dao.seance']->getFilmDAO()->findAllByCinemaIdNotIn($cinemaID);
-        }
-        // sinon, on retourne à l'accueil
-        else {
+            $filmsUnplanned = $app['dao.seance']->getFilmDAO()->findAllByCinemaIdNotIn($cinemaId);
+        } else {
             // renvoi à la page d'accueil
             return $app->redirect($request->getBasePath() . '/home');
         }
 
         // on récupère la liste des films de ce cinéma
-        $films   = $app['dao.seance']->getFilmDAO()->findAllByCinemaId($cinemaID);
+        $films = $app['dao.seance']->getFilmDAO()->findAllByCinemaId($cinemaId);
         // on récupère toutes les séances de films pour un cinéma donné
-        $seances = $app['dao.seance']->findAllByCinemaId($films, $cinemaID);
+        $seances = $app['dao.seance']->findAllByCinemaId($films, $cinemaId);
 
         // données de la vue
         $donnees = [
@@ -127,38 +110,43 @@ class ShowtimesController extends Controller {
      * @param string $cinemaId
      * @return RedirectResponse
      */
-    public function deleteShowtime(Request $request = null,
-            Application $app = null, string $filmId = null,
-            string $cinemaId = null): RedirectResponse {
+    public function deleteShowtime(
+        Request $request = null,
+        Application $app = null,
+        string $filmId = null,
+        string $cinemaId = null
+    ) {
         // si l'utilisateur n'est pas connecté ou sinon s'il n'est pas amdinistrateur
-        if (!$app['session']->get('user') || $app['session']->get('user')['username'] !==
-                'admin@adm.adm') {
-            // renvoi à la page d'accueil
-            return $app->redirect($request->getBasePath() . '/home');
-        }
+        $this->redirectIfUserNotConnectedOrNotAdmin($request, $app);
 
         // si la méthode de formulaire est la méthode POST
         if ($request->isMethod('POST')) {
 
             // on assainie les variables
-            $entries             = $this->extractArrayFromPostRequest($request,
-                    [
-                'heureDebut',
-                'heureFin',
-                'version',
-                'from']);
+            $entries = $this->extractArrayFromPostRequest(
+                $request,
+                [
+                    'heureDebut',
+                    'heureFin',
+                    'version',
+                    'from',
+                ]
+            );
             $entries['cinemaID'] = $cinemaId;
-            $entries['filmID']   = $filmId;
+            $entries['filmID'] = $filmId;
 
             // suppression de la séance
-            $app['dao.seance']->delete($entries['cinemaID'], $entries['filmID'],
-                    $entries['heureDebut'], $entries['heureFin']
+            $app['dao.seance']->delete(
+                $entries['cinemaID'],
+                $entries['filmID'],
+                $entries['heureDebut'],
+                $entries['heureFin'],
             );
             // en fonction d'où je viens, je redirige
-            if (strstr($entries['from'], 'movie')) {
-                return $app->redirect($request->getBasePath() . '/showtime/movie/' . $entries['filmID']);
+            if (strstr($entries['from'], 'movie') !== false) {
+                return $app->redirect($request->getBasePath() . '/showtime/movie/' . $filmId);
             } else {
-                return $app->redirect($request->getBasePath() . '/showtime/cinema/' . $entries['cinemaID']);
+                return $app->redirect($request->getBasePath() . '/showtime/cinema/' . $cinemaId);
             }
         } else {
             // renvoi à la page d'accueil
@@ -174,19 +162,18 @@ class ShowtimesController extends Controller {
      * @param string $cinemaId
      * @return string
      */
-    public function editShowtime(Request $request = null,
-            Application $app = null, string $filmId = null,
-            string $cinemaId = null) {
+    public function editShowtime(
+        Request $request = null,
+        Application $app = null,
+        string $filmId = null,
+        string $cinemaId = null
+    ) {
         // si l'utilisateur n'est pas connecté ou sinon s'il n'est pas amdinistrateur
-        if (!$app['session']->get('user') || $app['session']->get('user')['username'] !==
-                'admin@adm.adm') {
-            // renvoi à la page d'accueil
-            return $app->redirect($request->getBasePath() . '/home');
-        }
+        $this->redirectIfUserNotConnectedOrNotAdmin($request, $app);
 
         // init. des flags. Etat par défaut => je viens du cinéma et je créé
-        $fromCinema    = true;
-        $fromFilm      = false;
+        $fromCinema = true;
+        $fromFilm = false;
         $isItACreation = true;
         $alreadyExists = false;
 
@@ -194,40 +181,48 @@ class ShowtimesController extends Controller {
         $seanceOld = [
             'dateheureDebutOld' => '',
             'dateheureFinOld'   => '',
-            'heureFinOld'       => ''];
-
+            'heureFinOld'       => '',
+        ];
         $seance = null;
 
         // si l'on est en GET
-        if (filter_input(INPUT_SERVER, 'REQUEST_METHOD') == 'GET') {
+        if (filter_input(INPUT_SERVER, 'REQUEST_METHOD') === 'GET') {
             // on assainie les variables
-            $entries = $this->extractArrayFromGetRequest($request,
-                    [
-                'from',
-                'heureDebut',
-                'heureFin',
-                'version',
-                'filmID',
-                'cinemaID']);
+            $entries = $this->extractArrayFromGetRequest(
+                $request,
+                [
+                    'from',
+                    'heureDebut',
+                    'heureFin',
+                    'version',
+                    'filmID',
+                    'cinemaID',
+                ]
+            );
             // si le filmID n'était pas dans la requête GET
-            if (!$entries['filmID'] && $filmId) {
+            if ($entries['filmID'] === null && $filmId !== null) {
                 // il est dans la route
                 $entries['filmID'] = $filmId;
             }
             // si le cinemaID n'était pas dans la requête GET
-            if (!$entries['cinemaID'] && $cinemaId) {
+            if ($entries['cinemaID'] === null && $cinemaId !== null) {
                 // il est dans la route
                 $entries['cinemaID'] = $cinemaId;
             }
             // pour l'instant, on vérifie les données en GET
-            if ($entries && isset($entries['cinemaID'], $entries['filmID'],
-                            $entries['from'])) {
+            if (
+                isset(
+                    $entries['cinemaID'],
+                    $entries['filmID'],
+                    $entries['from']
+                ) === true
+            ) {
                 // on récupère l'identifiant du cinéma
                 $cinemaID = $entries['cinemaID'];
                 // l'identifiant du film
-                $filmID   = $entries['filmID'];
+                $filmID = $entries['filmID'];
                 // d'où vient on ?
-                $from     = $entries['from'];
+                $from = $entries['from'];
 
                 // puis on récupère les informations du cinéma en question
                 $cinema = $app['dao.seance']->getCinemaDAO()->find($cinemaID);
@@ -236,74 +231,98 @@ class ShowtimesController extends Controller {
                 $film = $app['dao.seance']->getFilmDAO()->find($filmID);
 
                 // s'il on vient des séances du film
-                if (strstr($entries['from'], 'movie')) {
+                if (strstr($from, 'movie') !== false) {
                     $fromCinema = false;
                     // on vient du film
-                    $fromFilm   = true;
+                    $fromFilm = true;
                 }
 
                 // ici, on veut savoir si on modifie ou si on ajoute
-                if (isset($entries['heureDebut'], $entries['heureFin'],
-                                $entries['version'])) {
+                if (
+                    isset(
+                        $entries['heureDebut'],
+                        $entries['heureFin'],
+                        $entries['version']
+                    ) === true
+                ) {
                     // nous sommes dans le cas d'une modification
-                    $isItACreation                  = false;
-                    $seance                         = new \Semeformation\Mvc\Cinema_crud\models\Seance();
+                    $isItACreation = false;
+                    $seance = new Seance();
                     // on récupère les anciennes valeurs (utile pour retrouver la séance avant de la modifier
                     $seanceOld['dateheureDebutOld'] = $entries['heureDebut'];
-                    $seanceOld['dateheureFinOld']   = $entries['heureFin'];
+                    $seanceOld['dateheureFinOld'] = $entries['heureFin'];
                     // dates PHP
-                    $dateheureDebut                 = new DateTime($entries['heureDebut']);
-                    $dateheureFin                   = new DateTime($entries['heureFin']);
+                    $dateheureDebut = new DateTime($entries['heureDebut']);
+                    $dateheureFin = new DateTime($entries['heureFin']);
                     // découpage en heures
                     $seance->setHeureDebut($dateheureDebut);
                     $seance->setHeureFin($dateheureFin);
                     // on récupère la version
                     $seance->setVersion($entries['version']);
                 }
-            }
-            // sinon, on retourne à l'accueil
-            else {
-                // renvoi à la page d'accueil
-                return $app->redirect($request->getBasePath() . '/home');
+            } else {
+                // en fonction d'où je viens, je redirige
+                return $this->redirectFrom($app, $request, $entries['from'], $entries['filmID'] ?? "", $entries['cinemaID'] ?? "");
             }
             // sinon, on est en POST
-        } else if (filter_input(INPUT_SERVER, 'REQUEST_METHOD') == 'POST') {
+        } elseif (filter_input(INPUT_SERVER, 'REQUEST_METHOD') === 'POST') {
             // on assainie les variables
-            $entries             = $this->extractArrayFromPostRequest($request,
-                    [
-                'datedebut',
-                'heuredebut',
-                'datefin',
-                'heurefin',
-                'dateheurefinOld',
-                'dateheuredebutOld',
-                'version',
-                'from',
-                'modificationInProgress']);
+            $entries = $this->extractArrayFromPostRequest(
+                $request,
+                [
+                    'datedebut',
+                    'heuredebut',
+                    'datefin',
+                    'heurefin',
+                    'dateheurefinOld',
+                    'dateheuredebutOld',
+                    'version',
+                    'from',
+                    'modificationInProgress',
+                ]
+            );
             $entries['cinemaID'] = $cinemaId;
-            $entries['filmID']   = $filmId;
+            $entries['filmID'] = $filmId;
             // d'où vient on ?
-            $from                = $entries['from'];
+            $from = $entries['from'];
             // si toutes les valeurs sont renseignées
-            if ($entries && isset($entries['cinemaID'], $entries['filmID'],
-                            $entries['datedebut'], $entries['heuredebut'],
-                            $entries['datefin'], $entries['heurefin'],
-                            $entries['dateheuredebutOld'],
-                            $entries['dateheurefinOld'], $entries['version'],
-                            $entries['from'])) {
+            if ($entries !== null
+                && isset(
+                    $entries['cinemaID'],
+                    $entries['filmID'],
+                    $entries['datedebut'],
+                    $entries['heuredebut'],
+                    $entries['datefin'],
+                    $entries['heurefin'],
+                    $entries['dateheuredebutOld'],
+                    $entries['dateheurefinOld'],
+                    $entries['version'],
+                    $entries['from']
+                ) === true
+            ) {
                 // nous sommes en Français
                 setlocale(LC_TIME, 'fra_fra');
                 // date du jour de projection de la séance
-                $datetimeDebut = DateTime::createFromFormat('d/m/Y H:i',
-                                $entries['datedebut'] . ' ' . $entries['heuredebut']);
-                $datetimeFin   = DateTime::createFromFormat('d/m/Y H:i',
-                                $entries['datefin'] . ' ' . $entries['heurefin']);
+                $datetimeDebut = DateTime::createFromFormat(
+                    'Y-m-d H:i',
+                    $entries['datedebut'] . ' ' . $entries['heuredebut']
+                );
+                $datetimeFin = DateTime::createFromFormat(
+                    'Y-m-d H:i',
+                    $entries['datefin'] . ' ' . $entries['heurefin']
+                );
+                if ($datetimeDebut === false) {
+                    $datetimeDebut = new DateTime();
+                }
+                if ($datetimeFin === false) {
+                    $datetimeFin = new DateTime();
+                }
                 // je récupère l'objet Cinema
-                $cinema        = $app['dao.seance']->getCinemaDAO()->find($entries['cinemaID']);
+                $cinema = $app['dao.seance']->getCinemaDAO()->find($entries['cinemaID']);
                 // je récupère l'objet Film
-                $film          = $app['dao.seance']->getFilmDAO()->find($entries['filmID']);
+                $film = $app['dao.seance']->getFilmDAO()->find($entries['filmID']);
                 // on crée l'objet Seance
-                $seance        = new \Semeformation\Mvc\Cinema_crud\models\Seance();
+                $seance = new Seance();
                 $seance->setHeureDebut($datetimeDebut);
                 $seance->setHeureFin($datetimeFin);
                 $seance->setVersion($entries['version']);
@@ -311,24 +330,18 @@ class ShowtimesController extends Controller {
                 $seance->setFilm($film);
                 try {
                     // je sauvegarde l'objet en BDD
-                    $app['dao.seance']->save($seance,
-                            $entries['dateheuredebutOld'],
-                            $entries['dateheurefinOld']);
+                    $app['dao.seance']->save(
+                        $seance,
+                        $entries['dateheuredebutOld'],
+                        $entries['dateheurefinOld']
+                    );
                     // en fonction d'où je viens, je redirige
-                    if (strstr($entries['from'], 'movie')) {
-                        // on redirige vers les séances du film
-                        return $app->redirect($request->getBasePath() . '/showtime/movie/' . $entries['filmID']);
-                    } else {
-                        // on redirige vers les séances du cinéma
-                        return $app->redirect($request->getBasePath() . '/showtime/cinema/' . $entries['cinemaID']);
-                    }
+                    return $this->redirectFrom($app, $request, $entries['from'], $entries['filmID'], $entries['cinemaID']);
                 } catch (BusinessObjectAlreadyExists $ex) {
                     $alreadyExists = true;
                 }
             }
-        }
-        // sinon, on retourne à l'accueil
-        else {
+        } else {
             // renvoi à la page d'accueil
             return $app->redirect($request->getBasePath() . '/home');
         }
@@ -349,4 +362,25 @@ class ShowtimesController extends Controller {
         return $app['twig']->render('showtime.edit.html.twig', $donnees);
     }
 
+    /**
+     * Redirige en fonction d'où l'utilisateur vient
+     *
+     * @param Application $app
+     * @param Request $request
+     * @param string $from
+     * @param string $filmId
+     * @param string $cinemaId
+     * @return RedirectResponse
+     */
+    private function redirectFrom(Application $app, Request $request, string $from, string $filmId = "", string $cinemaId = ""): RedirectResponse
+    {
+        // en fonction d'où je viens, je redirige
+        if (strstr($from, 'movie') !== false) {
+            // on redirige vers les séances du film
+            return $app->redirect($request->getBasePath() . '/showtime/movie/' . $filmId);
+        } else {
+            // on redirige vers les séances du cinéma
+            return $app->redirect($request->getBasePath() . '/showtime/cinema/' . $cinemaId);
+        }
+    }
 }
