@@ -11,48 +11,53 @@ use Semeformation\Mvc\Cinema_crud\models\Utilisateur;
  *
  * @author User
  */
-class Prefere {
-
+class Prefere
+{
     /**
      * Utilisateur
-     * @var \Semeformation\Mvc\Cinema_crud\models\Utilisateur 
+     * @var \Semeformation\Mvc\Cinema_crud\models\Utilisateur
      */
     private $utilisateur;
 
     /**
      * Film
-     * @var \Semeformation\Mvc\Cinema_crud\models\Film 
+     * @var \Semeformation\Mvc\Cinema_crud\models\Film
      */
     private $film;
 
     /**
      * Commentaire
-     * @var string 
+     * @var string
      */
     private $commentaire;
 
-    public function getUtilisateur() {
+    public function getUtilisateur(): Utilisateur
+    {
         return $this->utilisateur;
     }
 
-    public function getFilm() {
+    public function getFilm()
+    {
         return $this->film;
     }
 
-    public function setUtilisateur(Utilisateur $utilisateur) {
+    public function setUtilisateur(Utilisateur $utilisateur): void
+    {
         $this->utilisateur = $utilisateur;
     }
 
-    public function setFilm(Film $film) {
+    public function setFilm(Film $film): void
+    {
         $this->film = $film;
     }
 
-    public function getCommentaire() {
+    public function getCommentaire()
+    {
         return $this->commentaire;
     }
 
-    public function setCommentaire($commentaire) {
+    public function setCommentaire($commentaire): void
+    {
         $this->commentaire = $commentaire;
     }
-
 }
