@@ -2,52 +2,65 @@
 
 namespace Semeformation\Mvc\Cinema_crud\models;
 
-class Utilisateur {
-
+class Utilisateur
+{
     private $userId;
+
     private $nom;
+
     private $prenom;
+
     private $adresseCourriel;
+
     private $password;
 
-    public function getUserId() {
+    public function getUserId()
+    {
         return $this->userId;
     }
 
-    public function getNom() {
+    public function getNom()
+    {
         return $this->nom;
     }
 
-    public function getPrenom() {
+    public function getPrenom()
+    {
         return $this->prenom;
     }
 
-    public function getAdresseCourriel() {
+    public function getAdresseCourriel(): string
+    {
         return $this->adresseCourriel;
     }
 
-    public function getPassword() {
+    public function getPassword(): string
+    {
         return $this->password;
     }
 
-    public function setUserId($userId) {
+    public function setUserId($userId): void
+    {
         $this->userId = $userId;
     }
 
-    public function setNom($nom) {
+    public function setNom($nom): void
+    {
         $this->nom = $nom;
     }
 
-    public function setPrenom($prenom) {
+    public function setPrenom($prenom): void
+    {
         $this->prenom = $prenom;
     }
 
-    public function setAdresseCourriel($adresseCourriel) {
+    public function setAdresseCourriel($adresseCourriel): void
+    {
         $this->adresseCourriel = $adresseCourriel;
     }
 
-    public function setPassword($password) {
+    public function setPassword($password): void
+    {
         $this->password = $password;
     }
-
 }
